folder struct reorg
