--- src/api/v2/routes/support/search.js
@@ -1,17 +1,0 @@
-const validate = require('../../validators/support/search');
-
-module.exports = ({ instance, headers }) => {
-  const url = `https://${instance}.zendesk.com`;
-
-  return (options = {}) => {
-    const { error } = validate(options);
-    if (error) throw new Error(error.details[0].message);
-
-    const { search_string } = options;
-    return {
-      method: 'GET',
-      url: `${url}/api/v2/search.json?query=${search_string}`,
-      headers
-    };
-  };
-};
