--- src/v2.js
@@ -1,3 +1,0 @@
-module.exports = ({ instance, headers }) => ({
-  support: require('./api/v2/support')({ instance, headers })
-});
