+++ src/v2/index.js
@@ -1,0 +1,3 @@
+module.exports = ({ instance, headers }) => ({
+  support: require('./support')({ instance, headers })
+});
