+++ src/v2/support/ticket_metrics/validate.js
